api/requirements.txt
requests
azure-storage-blob
azure-functions